requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.37.3
    # via matplotlib
idna==3.4
    # via requests
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.0
    # via src (setup.py)
numpy==1.23.3
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   src (setup.py)
    #   torchvision
packaging==21.3
    # via matplotlib
pandas==1.5.0
    # via src (setup.py)
pillow==9.2.0
    # via
    #   matplotlib
    #   torchvision
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.2.1
    # via pandas
requests==2.28.1
    # via torchvision
six==1.16.0
    # via python-dateutil
torch==1.12.1
    # via
    #   src (setup.py)
    #   torchvision
torchvision==0.13.1
    # via src (setup.py)
typing-extensions==4.3.0
    # via
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
